Move middleware from startDevServer to own file
